Use opt-level=1 by default for dev.
As noted in the Makefile, the default debug mode (opt-level=0) is extraordinarily wasteful of stack space. We turn on a low level of optimisation in debug mode. According to LLVM documentation, O1 (equivalent to `rustc -C opt-level=1`) will "Optimize quickly without destroying debuggability." In practice this is usually safe, and solves our issue here.